+++ app/components/categoryManager/index.ts
@@ -1,0 +1,6 @@
+/*
+Copyright (C) 2024 Michael Lim - React Native Todo App - This software is free to use, modify, and share under the terms of the GNU General Public License v3.
+*/
+import { CategoryManager } from './categoryManager';
+
+export { CategoryManager };
